Move cronjob testdata to a v1 subdirectory
