+++ pkg/crd/testdata/cronjob/v1/cronjob_crdv1.yaml
