+++ pkg/crd/testdata/cronjob/v1/cronjob_types.go
@@ -16,11 +16,9 @@ limitations under the License.
 // TODO(directxman12): test this across both versions (right now we're just
 // trusting k/k conversion, which is probably fine though)
 
-//go:generate ../../../.run-controller-gen.sh crd:crdVersions=v1 paths=. output:dir=.
-
 // +groupName=testdata.kubebuilder.io
 // +versionName=v1
-package cronjob
+package v1
 
 import (
 	"fmt"
